Store PostgreSQL data in host file system
The data is stored in a subdirectory. Reason (from the README of the Docker image):

If the data volume you're using is a filesystem mountpoint (like with GCE persistent disks) or remote folder that cannot be chowned to the postgres user (like some NFS mounts), Postgres initdb recommends a subdirectory be created to contain the data.